+++ e2e/transform/multiple-transformers/filePreprocessor.js
@@ -8,7 +8,7 @@
 const path = require('path');
 
 module.exports = {
-  process(src, filename, config, options) {
+  process(src, filename) {
     return `
       module.exports = '${path.basename(filename)}';
     `;