+++ e2e/transform/transform-testrunner/__tests__/add.test.js
@@ -1,4 +1,3 @@
-#!/usr/bin/env node
 /**
  * Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
  *
@@ -6,8 +5,6 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-if (process.env.NODE_ENV == null) {
-  process.env.NODE_ENV = 'test';
-}
-
-require('../build/cli').run();
+it('should add two numbers', () => {
+  expect(1 + 1).toBe(2);
+});