+++ packages/jest-core/src/FailedTestsCache.ts
@@ -6,7 +6,6 @@
  */
 
 import type {TestResult} from '@jest/test-result';
-import type {Config} from '@jest/types';
 import type {Test} from 'jest-runner';
 
 type TestMap = Record<string, Record<string, boolean>>;
@@ -37,14 +36,5 @@ export default class FailedTestsCache {
       }, {});
 
     this._enabledTestsMap = Object.freeze(this._enabledTestsMap);
-  }
-
-  updateConfig(globalConfig: Config.GlobalConfig): Config.GlobalConfig {
-    if (!this._enabledTestsMap) {
-      return globalConfig;
-    }
-    const newConfig: Config.GlobalConfig = {...globalConfig};
-    newConfig.enabledTestsMap = this._enabledTestsMap;
-    return Object.freeze(newConfig);
   }
 }