+++ packages/jest-create-cache-key-function/src/index.ts
@@ -12,17 +12,33 @@ import {readFileSync} from 'fs';
 import {relative} from 'path';
 import type {Config} from '@jest/types';
 
-type CacheKeyOptions = {
+type OldCacheKeyOptions = {
   config: Config.ProjectConfig;
   instrument: boolean;
 };
 
-type GetCacheKeyFunction = (
+// Should mirror `import('@jest/transform').TransformOptions`
+type NewCacheKeyOptions = {
+  config: Config.ProjectConfig;
+  configString: string;
+  instrument: boolean;
+};
+
+type OldGetCacheKeyFunction = (
   fileData: string,
   filePath: Config.Path,
   configStr: string,
-  options: CacheKeyOptions,
+  options: OldCacheKeyOptions,
 ) => string;
+
+// Should mirror `import('@jest/transform').Transformer['getCacheKey']`
+type NewGetCacheKeyFunction = (
+  sourceText: string,
+  sourcePath: Config.Path,
+  options: NewCacheKeyOptions,
+) => string;
+
+type GetCacheKeyFunction = OldGetCacheKeyFunction | NewGetCacheKeyFunction;
 
 function getGlobalCacheKey(files: Array<string>, values: Array<string>) {
   return [
@@ -39,14 +55,18 @@ function getGlobalCacheKey(files: Array<string>, values: Array<string>) {
 }
 
 function getCacheKeyFunction(globalCacheKey: string): GetCacheKeyFunction {
-  return (src, file, _configString, options) => {
-    const {config, instrument} = options;
+  return (sourceText, sourcePath, configString, options) => {
+    // Jest 27 passes a single options bag which contains `configString` rather than as a separate argument.
+    // We can hide that API difference, though, so this module is usable for both jest@<27 and jest@>=27
+    const inferredOptions = options || configString;
+    const {config, instrument} = inferredOptions;
+
     return createHash('md5')
       .update(globalCacheKey)
       .update('\0', 'utf8')
-      .update(src)
+      .update(sourceText)
       .update('\0', 'utf8')
-      .update(config.rootDir ? relative(config.rootDir, file) : '')
+      .update(config.rootDir ? relative(config.rootDir, sourcePath) : '')
       .update('\0', 'utf8')
       .update(instrument ? 'instrument' : '')
       .digest('hex');