+++ packages/jest-message-util/src/index.ts
@@ -26,8 +26,7 @@ const stackUtils = new StackUtils({cwd: 'something which does not exist'});
 let nodeInternals: Array<RegExp> = [];
 
 try {
-  // https://github.com/tapjs/stack-utils/issues/54
-  nodeInternals = StackUtils.nodeInternals().concat(/\s*\(node:/);
+  nodeInternals = StackUtils.nodeInternals();
 } catch {
   // `StackUtils.nodeInternals()` fails in browsers. We don't need to remove
   // node internals in the browser though, so no issue.