+++ packages/jest-repl/src/cli/args.ts
@@ -7,12 +7,45 @@
  */
 
 import type {Options} from 'yargs';
-import Runtime from 'jest-runtime';
 
 export const usage = 'Usage: $0 [--config=<pathToConfigFile>]';
 
+const runtimeCLIOptions: Record<
+  'cache' | 'config' | 'debug' | 'version' | 'watchman',
+  Options
+> = {
+  cache: {
+    default: true,
+    description:
+      'Whether to use the preprocessor cache. Disable ' +
+      'the cache using --no-cache.',
+    type: 'boolean',
+  },
+  config: {
+    alias: 'c',
+    description: 'The path to a Jest config file.',
+    type: 'string',
+  },
+  debug: {
+    description: 'Print debugging info about your jest config.',
+    type: 'boolean',
+  },
+  version: {
+    alias: 'v',
+    description: 'Print the version and exit',
+    type: 'boolean',
+  },
+  watchman: {
+    default: true,
+    description:
+      'Whether to use watchman for file crawling. Disable using ' +
+      '--no-watchman.',
+    type: 'boolean',
+  },
+};
+
 export const options: Record<string, Options> = {
-  ...Runtime.getCLIOptions(),
+  ...runtimeCLIOptions,
   replname: {
     alias: 'r',
     description: